CSI-Mario-Sanchez
Print changed to one line with \n line breaks and an additional file was added to import the print method into another file.
