--- Projects/ACSII-initials.py
@@ -1,42 +1,0 @@
-def print_A():
-    print("""    A    
-   / \   
-  /___\  
- /     \ 
-/       \\""")
-def print_I():
-    print(""" _______ 
-    |    
-    |    
-    |    
- ___|___ """)
-def print_M():
-    print("""|\     /|
-| \   / |
-|  \ /  |
-|   V   |
-|       |""")
-def print_O():
-    print("""  _____  
- /     \ 
- |     | 
- |     | 
- \_____/ """)
-def print_R():
-    print("""  ____  
-  |   \ 
-  |___/  
-  | \   
-  |  \  
-  |   \  """)
-print_M()
-print()
-print_A()
-print()
-print_R()
-print()
-print_I()
-print()
-print_O()
-print()
-name = input("")
