+++ Projects/ASCII_Initials.py
@@ -1,0 +1,23 @@
+def print_A():
+    print("    A    \n   / \   \n  /___\  \n /     \ \n/       \\")
+def print_I():
+    print(" _______ \n    |    \n    |    \n    |    \n ___|___ ")
+def print_M():
+    print("|\     /|\n| \   / |\n|  \ /  |\n|   V   |\n|       |")
+def print_O():
+    print("  _____  \n /     \ \n |     | \n |     | \n \_____/ ")
+def print_R():
+    print("  ____  \n  |   \ \n  |___/  \n  | \   \n  |  \  \n  |   \  ")
+def print_name():
+    print_M()
+    print()
+    print_A()
+    print()
+    print_R()
+    print()
+    print_I()
+    print()
+    print_O()
+    print()
+print_name()
+test = input();
