+++ Projects/ASCII_Name_Print.py
@@ -1,0 +1,5 @@
+import sys
+import ASCII_Initials
+
+print_name()
+test = input();
